+++ runtime/integration_edit_test.go
@@ -84,6 +84,6 @@ func TestMouseLeftClick(t *testing.T) {
 	te.SetCursor(1, 0)
 
 	gutter := te.WP().GutterWidth()
-	te.Click(0, uint32(gutter+5))
+	te.Click(0, gutter+5)
 	te.ExpectCursor(1, 5)
 }
